Resize input based on word length
this is a big & non-atomic commit, but iirc it's forbidden to
`git add --patch` on the sabbath

I added a hidden element with the same text in it as the input and the same
formatting, and use DOM inspection to get its width every time the word is
changed.
I use the word "size" instead of "width" in the code bc "width" is the hardest
word in the english language for me to dictate reliably, so I end up having to
spell it out every time.
I changed around styles a bit as well (rip rounded corners).